test fixes + import fixes
