+++ packages/ai/src/generate-object/stream-object.test.ts
@@ -19,8 +19,6 @@ import { verifyNoObjectGeneratedError } from '../error/verify-no-object-generate
 import * as logWarningsModule from '../logger/log-warnings';
 import { MockLanguageModelV4 } from '../test/mock-language-model-v4';
 import { createMockServerResponse } from '../test/mock-server-response';
-import { MockTracer } from '../test/mock-tracer';
-import { OpenTelemetryIntegration } from '../telemetry/open-telemetry-integration';
 import { AsyncIterableStream } from '../util/async-iterable-stream';
 import { streamObject } from './stream-object';
 import { StreamObjectResult } from './stream-object-result';
@@ -1406,193 +1404,6 @@ describe('streamObject', () => {
           "type": "json",
         }
       `);
-    });
-  });
-
-  describe('telemetry', () => {
-    let tracer: MockTracer;
-
-    beforeEach(() => {
-      tracer = new MockTracer();
-    });
-
-    it('should not record any telemetry data when not explicitly enabled', async () => {
-      const result = streamObject({
-        model: new MockLanguageModelV4({
-          doStream: async () => ({
-            stream: convertArrayToReadableStream([
-              {
-                type: 'response-metadata',
-                id: 'id-0',
-                modelId: 'mock-model-id',
-                timestamp: new Date(0),
-              },
-              { type: 'text-start', id: '1' },
-              { type: 'text-delta', id: '1', delta: '{ ' },
-              { type: 'text-delta', id: '1', delta: '"content": ' },
-              { type: 'text-delta', id: '1', delta: `"Hello, ` },
-              { type: 'text-delta', id: '1', delta: `world` },
-              { type: 'text-delta', id: '1', delta: `!"` },
-              { type: 'text-delta', id: '1', delta: ' }' },
-              { type: 'text-end', id: '1' },
-              {
-                type: 'finish',
-                finishReason: { unified: 'stop', raw: 'stop' },
-                usage: testUsage,
-              },
-            ]),
-          }),
-        }),
-        schema: z.object({ content: z.string() }),
-        prompt: 'prompt',
-        experimental_telemetry: {
-          integrations: new OpenTelemetryIntegration({ tracer }),
-        },
-        _internal: { now: () => 0 },
-      });
-
-      // consume stream
-      await convertAsyncIterableToArray(result.partialObjectStream);
-
-      expect(tracer.jsonSpans).toMatchSnapshot();
-    });
-
-    it('should record telemetry data when enabled', async () => {
-      const result = streamObject({
-        model: createTestModel({
-          stream: convertArrayToReadableStream([
-            {
-              type: 'response-metadata',
-              id: 'id-0',
-              modelId: 'mock-model-id',
-              timestamp: new Date(0),
-            },
-            { type: 'text-start', id: '1' },
-            { type: 'text-delta', id: '1', delta: '{ ' },
-            { type: 'text-delta', id: '1', delta: '"content": ' },
-            { type: 'text-delta', id: '1', delta: `"Hello, ` },
-            { type: 'text-delta', id: '1', delta: `world` },
-            { type: 'text-delta', id: '1', delta: `!"` },
-            { type: 'text-delta', id: '1', delta: ' }' },
-            { type: 'text-end', id: '1' },
-            {
-              type: 'finish',
-              finishReason: { unified: 'stop', raw: 'stop' },
-              usage: testUsage,
-              providerMetadata: {
-                testProvider: {
-                  testKey: 'testValue',
-                },
-              },
-            },
-          ]),
-        }),
-        schema: z.object({ content: z.string() }),
-        schemaName: 'test-name',
-        schemaDescription: 'test description',
-        prompt: 'prompt',
-        topK: 0.1,
-        topP: 0.2,
-        frequencyPenalty: 0.3,
-        presencePenalty: 0.4,
-        temperature: 0.5,
-        headers: {
-          header1: 'value1',
-          header2: 'value2',
-        },
-        experimental_telemetry: {
-          isEnabled: true,
-          functionId: 'test-function-id',
-          metadata: {
-            test1: 'value1',
-            test2: false,
-          },
-          integrations: new OpenTelemetryIntegration({ tracer }),
-        },
-        _internal: { now: () => 0 },
-      });
-
-      // consume stream
-      await convertAsyncIterableToArray(result.partialObjectStream);
-
-      expect(tracer.jsonSpans).toMatchSnapshot();
-    });
-
-    it('should not record telemetry inputs / outputs when disabled', async () => {
-      const result = streamObject({
-        model: new MockLanguageModelV4({
-          doStream: async () => ({
-            stream: convertArrayToReadableStream([
-              {
-                type: 'response-metadata',
-                id: 'id-0',
-                modelId: 'mock-model-id',
-                timestamp: new Date(0),
-              },
-              { type: 'text-start', id: '1' },
-              { type: 'text-delta', id: '1', delta: '{ ' },
-              { type: 'text-delta', id: '1', delta: '"content": ' },
-              { type: 'text-delta', id: '1', delta: `"Hello, ` },
-              { type: 'text-delta', id: '1', delta: `world` },
-              { type: 'text-delta', id: '1', delta: `!"` },
-              { type: 'text-delta', id: '1', delta: ' }' },
-              { type: 'text-end', id: '1' },
-              {
-                type: 'finish',
-                finishReason: { unified: 'stop', raw: 'stop' },
-                usage: testUsage,
-              },
-            ]),
-          }),
-        }),
-        schema: z.object({ content: z.string() }),
-        prompt: 'prompt',
-        experimental_telemetry: {
-          isEnabled: true,
-          recordInputs: false,
-          recordOutputs: false,
-          integrations: new OpenTelemetryIntegration({ tracer }),
-        },
-        _internal: { now: () => 0 },
-      });
-
-      // consume stream
-      await convertAsyncIterableToArray(result.partialObjectStream);
-
-      expect(tracer.jsonSpans).toMatchSnapshot();
-    });
-
-    it('should pass correct callId to onError when doStream throws', async () => {
-      const onErrorCalls: Array<{ callId: string; error: unknown }> = [];
-
-      const result = streamObject({
-        model: new MockLanguageModelV4({
-          doStream: async () => {
-            throw new Error('doStream failure');
-          },
-        }),
-        schema: z.object({ content: z.string() }),
-        prompt: 'prompt',
-        experimental_telemetry: {
-          isEnabled: true,
-          integrations: {
-            onStart(event: any) {
-              onErrorCalls.length; // keep reference
-            },
-            onError(event: any) {
-              onErrorCalls.push(event);
-            },
-          },
-        },
-        onError: () => {},
-      });
-
-      await convertAsyncIterableToArray(result.partialObjectStream);
-
-      expect(onErrorCalls).toHaveLength(1);
-      expect(onErrorCalls[0].callId).not.toBe('');
-      expect(onErrorCalls[0].error).toBeInstanceOf(Error);
-      expect((onErrorCalls[0].error as Error).message).toBe('doStream failure');
     });
   });
